test-requirements.txt
molecule
molecule-docker
docker
ansible-lint
yamllint
flake8
pytest
pytest-testinfra
tox
# jmespath
